+++ DatadogTrace/Sources/Objc/Tracing/Trace+objc.swift
@@ -106,6 +106,10 @@ public final class objc_Trace: NSObject {
     public static func enable(with configuration: objc_TraceConfiguration) {
         Trace.enable(with: configuration.swiftConfig)
     }
+
+    public static func enable(with configuration: objc_TraceConfiguration, instanceName: String?) {
+        Trace.enable(with: configuration.swiftConfig, in: CoreRegistry.instance(named: instanceName ?? CoreRegistry.defaultInstanceName))
+    }
 }
 
 @objc(DDTracer)
@@ -124,6 +128,10 @@ public final class objc_Tracer: NSObject, objc_OTTracer {
 
     public static func shared() -> objc_OTTracer {
         objc_Tracer(swiftTracer: Tracer.shared())
+    }
+
+    public static func shared(instanceName: String?) -> objc_OTTracer {
+        objc_Tracer(swiftTracer: Tracer.shared(in: CoreRegistry.instance(named: instanceName ?? CoreRegistry.defaultInstanceName)))
     }
 
     public func startSpan(_ operationName: String) -> objc_OTSpan {