+++ DatadogRUM/Sources/Feature/RUMFeature.swift
@@ -240,6 +240,7 @@ internal final class RUMFeature: DatadogRemoteFeature, RUMSessionSamplerProvider
             uiKitRUMActionsPredicate: configuration.uiKitActionsPredicate,
             swiftUIRUMViewsPredicate: configuration.swiftUIViewsPredicate,
             swiftUIRUMActionsPredicate: configuration.swiftUIActionsPredicate,
+            scrollAndSwipeActionsDisabled: configuration.featureFlags[.disableScrollAndSwipeActions],
             longTaskThreshold: configuration.longTaskThreshold,
             appHangThreshold: configuration.appHangThreshold,
             mainQueue: configuration.mainQueue,